Fixing GL ES problems + cleanup parts

+++ Assets/Shaders/character.fs
@@ -1,4 +1,3 @@
-
 out vec4 FragColor;
 
 in vec2 aTexCoord;
@@ -8,9 +7,9 @@ uniform sampler2D textureA;
 
 uniform int textureID[864];
 
-void main() {	
+void main() {
 
-	vec2 pTexCoord = vec2(aTexCoord.x + (textureID[instanceID] % 16) / 16.0, aTexCoord.y - (textureID[instanceID] / 16) / 4.0);
+	vec2 pTexCoord = vec2(aTexCoord.x + float(textureID[instanceID] % 16) / 16.0, aTexCoord.y - float(textureID[instanceID] / 16) / 4.0);
 
 	FragColor = texture(textureA, pTexCoord);
 }
